split blocks
