+++ lib/asciidoctor/rfc/v3/converter.rb
@@ -1,6 +1,7 @@
 require "asciidoctor/rfc/common/base"
 require "asciidoctor/rfc/v3/front"
 require "asciidoctor/rfc/v3/lists"
+require "asciidoctor/rfc/v3/blocks"
 
 module Asciidoctor
   module RFC::V3
@@ -17,6 +18,7 @@ module Asciidoctor
       include ::Asciidoctor::RFC::Common::Base
       include ::Asciidoctor::RFC::V3::Front
       include ::Asciidoctor::RFC::V3::Lists
+      include ::Asciidoctor::RFC::V3::Blocks
 
       register_for "rfc3"
 
@@ -145,28 +147,6 @@ module Asciidoctor
 
       # Syntax:
       #   [[id]]
-      #   [align=left|center|right,alt=alt_text] (optional)
-      #   ....
-      #     literal
-      #   ....
-      def literal(node)
-        result = []
-        result << "<figure>" if node.parent.context != :example
-        id = set_header_attribute "anchor", node.id
-        align = get_header_attribute node, "align"
-        alt = set_header_attribute "alt", node.alt
-        type = set_header_attribute "type", "ascii-art"
-        result << "<artwork#{id}#{align}#{alt}#{type}>"
-        node.lines.each do |line|
-          result << line.gsub(/\&/, "&amp;").gsub(/</, "&lt;").gsub(/>/, "&gt;")
-        end
-        result << "</artwork>"
-        result << "</figure>" if node.parent.context != :example
-        result
-      end
-
-      # Syntax:
-      #   [[id]]
       #   [keepWithNext=true,keepWithPrevious=true] (optional)
       #   Text
       def paragraph(node)
@@ -179,69 +159,6 @@ module Asciidoctor
         keepWithNext = get_header_attribute node, "keepWithNext"
         keepWithPrevious = get_header_attribute node, "keepWithPrevious"
         result << "<t#{id}#{keepWithNext}#{keepWithPrevious}>#{node.content}</t>"
-        result
-      end
-
-      # Syntax:
-      #   [[id]]
-      #   [quote, attribution, citation info] # citation info limited to URL
-      #   Text
-      def quote(node)
-        result = []
-        id = set_header_attribute "anchor", node.id
-        quotedFrom = set_header_attribute "quotedFrom", node.attr("attribution")
-        citationInfo = node.attr "citetitle"
-        if !citationInfo.nil? && citationInfo =~ URI::DEFAULT_PARSER.make_regexp
-          cite = set_header_attribute "cite", citationInfo
-        end
-        result << "<blockquote#{id}#{quotedFrom}#{cite}>"
-        result << node.content
-        result << "</blockquote>"
-        result
-      end
-
-      # Syntax:
-      #   = Title
-      #   Author
-      #   :HEADER
-      #
-      #   ABSTRACT
-      #
-      #   NOTE: note
-      #
-      #   [NOTE]
-      #   .Title (in preamble)
-      #   ====
-      #     Content
-      #   ====
-      #
-      #     [NOTE,removeInRFC=true] (in preamble)
-      #     [NOTE,display=true|false,source=name] (in body)
-      #   .Title
-      #   ====
-      #     Content
-      #   ====
-      # @note admonitions within preamble are notes. Elsewhere, they are comments.
-      def admonition(node)
-        result = []
-        if node.parent.context == :preamble
-          if $seen_abstract
-            $seen_abstract = false
-            result << "</abstract>"
-          end
-          removeInRFC = get_header_attribute node, "removeInRFC"
-          result << "<note#{removeInRFC}>"
-          result << "<name>#{node.title}</name>" unless node.title.nil?
-          result << (paragraph1 node)
-          result << "</note>"
-        else
-          id = set_header_attribute "anchor", node.id
-          display = get_header_attribute node, "display"
-          source = get_header_attribute node, "source"
-          result << "<cref#{id}#{display}#{source}>"
-          result << node.content
-          result << "</cref>"
-        end
         result
       end
 
@@ -340,41 +257,6 @@ module Asciidoctor
 
       # Syntax:
       #   [[id]]
-      #   ****
-      #   Sidebar
-      #   ****
-      def sidebar(node)
-        result = []
-        id = set_header_attribute "anchor", node.id
-        result << "<aside#{id}>"
-        result << cell.content
-        result << "</aside>"
-        result
-      end
-
-      # Syntax:
-      #   .Title
-      #   ====
-      #   Example
-      #   ====
-      def example(node)
-        result = []
-        id = set_header_attribute "anchor", node.id
-        result << "<figure#{id}>"
-        result << %(<name>#{node.title}</name>) if node.title?
-        # TODO iref
-        result << node.content
-        result << "</figure>"
-        result.blocks.each do |b|
-          unless b == :listing or b == :image or b == :literal
-            warn "asciidoctor: WARNING: examples (figures) should only contain listings (sourcecode), images (artwork), or literal (artwork):\n#{b.text}"
-          end
-        end
-        result
-      end
-
-      # Syntax:
-      #   [[id]]
       #   .Name
       #   [link=xxx,align=left|center|right,alt=alt_text,type]
       #   image::filename[]
@@ -390,34 +272,6 @@ module Asciidoctor
         type = set_header_attribute node, "type", link =~ /\.svg$/ ? "svg" : "binary-art"
         name = nil
         result << "<artwork#{id}#{name}#{align}#{alt}#{type}#{src}/>"
-        result << "</figure>" if node.parent.context != :example
-        result
-      end
-
-      # Syntax:
-      #   .name
-      #   [source,type,src=uri] (src is mutually exclusive with listing content) (v3)
-      #   [source,type,src=uri,align,alt] (src is mutually exclusive with listing content) (v2)
-      #   ----
-      #   code
-      #   ----
-      def listing(node)
-        result = []
-        result << "<figure>" if node.parent.context != :example
-        align = nil
-        alt = nil
-        tag = "sourcecode"
-        id = set_header_attribute "anchor", node.id
-        name = set_header_attribute "name", node.title
-        type = set_header_attribute "type", node.attr("language")
-        src = set_header_attribute "src", node.attr("src")
-        result << "<#{tag}#{id}#{align}#{name}#{type}#{src}#{alt}>"
-        if src.nil?
-          node.lines.each do |line|
-            result << line.gsub(/\&/, "&amp;").gsub(/</, "&lt;").gsub(/>/, "&gt;")
-          end
-        end
-        result << "</#{tag}>"
         result << "</figure>" if node.parent.context != :example
         result
       end
